requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
paho-mqtt==1.6.1
requests==2.28.2
sqlalchemy==1.4.46
